requirements.txt
flask
flask-mysqldb
flask-login
flask-wtf
flask-bootstrap
mysqlclient 